TaiLieu/TAPIT/Python-va-Serial/4.code/requirements.txt
PyQt5
matplotlib
pyserial
pyinstaller